Revert "add thread logging"
This reverts commit 9dc22401054b8f91f2b2aa3033e482996913febb.
