+++ logger/src/lib.rs
@@ -2,10 +2,7 @@
 
 use {
     lazy_static::lazy_static,
-    std::{
-        io::Write,
-        sync::{Arc, RwLock},
-    },
+    std::sync::{Arc, RwLock},
 };
 
 lazy_static! {
@@ -55,25 +52,6 @@ pub fn setup_with_default(filter: &str) {
 // Configures logging with the default filter "error" if RUST_LOG is not set
 pub fn setup() {
     setup_with_default("error");
-}
-
-pub fn setup_logging_with_thread_id(filter: &str) {
-    let logger = env_logger::Builder::from_env(env_logger::Env::new().default_filter_or(filter))
-        .format(|buf, record| {
-            writeln!(
-                buf,
-                "{:?} [{} {} {}:{}] {}",
-                std::thread::current().id(),
-                buf.timestamp_nanos(),
-                buf.default_level_style(record.level())
-                    .value(record.level()),
-                record.target(),
-                record.line().unwrap_or(0),
-                record.args()
-            )
-        })
-        .build();
-    replace_logger(logger);
 }
 
 // Configures file logging with a default filter if RUST_LOG is not set
